updates to smart contract

+++ smart_contract/contracts/Transactions.sol
@@ -1,20 +1,17 @@
 // SPDX-License-Identifier: UNLICENSED
-
 pragma solidity ^0.8.0;
-
-import "hardhat/console.sol";
 
 contract Transactions {
     uint256 transactionCount;
 
-    event Transfer(address from, address receiver, uint amount, string message, uint256 timestamp, string keyword);
-  
+    event Transfer(address from, address receiver, uint amount, string message, uint timestamp, string keyword);
+
     struct TransferStruct {
         address sender;
         address receiver;
         uint amount;
         string message;
-        uint256 timestamp;
+        uint timestamp;
         string keyword;
     }
